added storing locations to database, added uploading locations after stopping tracking to remote database. Added metrics like duration and Distance

+++ app/src/main/java/com/markosopcic/cycler/data/dao/LocationDAO.kt
@@ -13,7 +13,7 @@ interface LocationDAO {
         fun AddLocation(location : Location)
 
         @Query("select * from Location where eventId = :id")
-        fun GetLocationsForEvent(id: Int) : List<Location>
+        fun GetLocationsForEvent(id: Long) : List<Location>
 
         @Delete
         fun RemoveLocations(locations : List<Location>)
